refactor: :metal: component tsx and types
